tweaking behaviours and fixes

+++ editor/src/tools/paint_tool.hpp
@@ -4,7 +4,7 @@
 namespace pf::editor {
 class PaintTool {
 public:
-    void on_mouse_down(int wx, int wy, EditorContext& ctx);
-    void on_mouse_drag(int wx, int wy, EditorContext& ctx);
+    void on_mouse_down(float wx, float wy, EditorContext& ctx);
+    void on_mouse_drag(float wx, float wy, EditorContext& ctx);
 };
 } // namespace pf::editor
